refactor(bulk file bundles): Move bundle limits to constants file PE-460
