+++ src/utils/constants.ts
@@ -32,16 +32,17 @@ export const fakeTxID = TxID('0000000000000000000000000000000000000000000');
  */
 export const minArDriveCommunityARTip = 0.000_010_000_000;
 
+/** These limits are being chosen as a precaution due to potential gateway limitations */
+export const MAX_BUNDLE_SIZE = 524_288_000; // 500 MiB
+export const MAX_DATA_ITEM_LIMIT = 500; // 500 data items
+
 // ArDrive Profit Sharing Community Smart Contract
 export const communityTxId = '-8A6RexFkpfWwuyVO98wzSFZh0d6VJuI-buTJvlwOJQ';
-
-/** Estimated AR price for most metadata transactions */
-export const assumedMetadataTxARPrice = 0.000_002_500_000;
 
 export const publicJsonContentTypeTag = { name: 'Content-Type', value: JSON_CONTENT_TYPE };
 export const privateOctetContentTypeTag = { name: 'Content-Type', value: PRIVATE_CONTENT_TYPE };
 
-export const privateCipherTag = { name: 'Cipher', value: 'AES256-GCM' };
+export const privateCipherTag = { name: 'Cipher', value: defaultCipher };
 export const fakePrivateCipherIVTag = { name: 'Cipher-IV', value: 'qwertyuiopasdfgh' }; // Cipher-IV is always 16 characters
 
 export const authTagLength = 16;
